requirements.txt
mne>=1.11.0
numpy>=2.4.3
matplotlib>=3.10.8
mysql-connector-python>=9.6.0
scipy>=1.17.1
